[8221] Refactor survey services to use a base class
